+++ src/app/services/auth.model.ts
@@ -1,0 +1,7 @@
+export interface GuestAuth {
+  success: boolean;
+  token: string;
+  user_id: string;
+  token_data: any;
+  nonce: string;
+}