+++ webpack.config.ts
@@ -3,3 +3,14 @@ import Dotenv from 'dotenv-webpack';
 module.exports = {
   plugins: [new Dotenv()],
 };
+
+/**
+ * @proxy
+ * 
+ * add this to angular.json under architect.serve
+ * 
+ * "options": {
+      "proxyConfig": "proxy.conf.json"
+    },
+ * 
+ */